backend/requirements.txt
flask
flask-cors
resquests
jonsify
pymongo
flask-mongoengine
flask-restful
dnspython
bson